* update gitignore for modern IntelliJ

* change artificat ID for beta

* Bandit random action placeholder (#1)

* ability to pass in assignment options and attributes

* simple test case

* placeholder bandit algorithm in place

* small changes from self-review

* more helpful RAC read failure message

* WIP wiring up logging

* log bandit info as well

* preserve bandit test data

* feedback from PR and adjust indention to match rest of project

* use Math.abs() as its more readable

* trailing newlines

* Bandit as a dynamic variant (#2)

* use dynamic variations

* separate bandit logger

* ability to log non-bandit selected control

* log variation not bandit boolean

* remove now unused class

* feedback from PR

* Deserialize and store bandit parameters (#3)

* deserialize bandit parameters

* cleaned up

* updated to break out subject and action attributes

* genericize configuration requestor

* changes from self-reivew of PR

* Falcon model for bandit parameter evaluation (#4)

* work in progress

* math in place

* working on tests still

* tests passing

* new test cases

* wip manually verifying math

* tests passing

* remove logging

* clean up score computation

* changes from self-reivew of PR

* consolidate test bandit rac setup code

* clean up test file

* update comments

* more comment improvements

* feedback from PR; mainly cleanup and adding some tuning parameters

* Have bandit logger separate out context (#5)

* work in progress

* math in place

* working on tests still

* tests passing

* new test cases

* wip manually verifying math

* tests passing

* remove logging

* clean up score computation

* changes from self-reivew of PR

* consolidate test bandit rac setup code

* clean up test file

* update comments

* more comment improvements

* failing tests in place

* break out attributes by type

* changes from self-review of PR

* return type as provided

* make EppoValue constructor private

* feedback from PR

* more fault tolerant reading of properties file (#6)

* catch all exceptions; unit test (#7)

* Rename BANDIT algorithm type to CONTEXTUAL_BANDIT to match Eppo (#8)

* catch all exceptions; unit test

* conform bandit algorithm type

* Add test for cold start (#9)

* explicit test for cold start

* upgrade gcloud to appease linter

* update deserializer test

* Convenience methods to serialize attributes to JSON string and other cleanup (#10)

* add back in isXXX checks

* helper function for JSON string serialization

* rename variation to banditKey for bandit logging

* remove accidentally added junit

* Prepare beta SDK for merge into Main one (#11)

* add graceful mode by default that does not throw exceptions (FF-949) (#29)

* add graceful mode by default that does not throw exceptions (FF-949)

* test graceful mode on and off

* test all functions

* remove from javadoc

* Bump com.github.tomakehurst:wiremock-jre8 from 2.33.2 to 2.35.1 (#28)

Bumps [com.github.tomakehurst:wiremock-jre8](https://github.com/wiremock/wiremock) from 2.33.2 to 2.35.1.
- [Release notes](https://github.com/wiremock/wiremock/releases)
- [Commits](https://github.com/wiremock/wiremock/compare/2.33.2...2.35.1)

---
updated-dependencies:
- dependency-name: com.github.tomakehurst:wiremock-jre8
  dependency-type: direct:development
...


* add compatibility for java 8 (#35)

* java8 compatibility using apache httpclient

* cleanup

* added timeouts

* add maven-gpg-plugin

* test across java 8, 11, 17

* pom

* remove List.of

* fix test

* eppo value

* fix plugin

* profile

---------


* Bump org.apache.httpcomponents:httpclient from 4.5.10 to 4.5.13 (#36)

Bumps org.apache.httpcomponents:httpclient from 4.5.10 to 4.5.13.

---
updated-dependencies:
- dependency-name: org.apache.httpcomponents:httpclient
  dependency-type: direct:production
...


* [EN-17858] FetchConfigurationsTask is more fault tolerant (#5) (#32)

* [EN-17858] FetchConfigurationsTask is more fault tolerant

* revert change to pom.xml, thereby having a single, clean commit that can be merged back to the Eppo main repo

* bump to version 2.2.0 (#38)

* add support for semver rule evaluation (FF-1569) (#39)

* add support for semver rule evaluation (FF-1569)

* azul builds

* adjust unit tests for semver comparison

* numeric first

* add back in isXXX checks

* helper function for JSON string serialization

* rename variation to banditKey for bandit logging

* remove accidentally added junit

---------


* remove duplicate test dependency

---------

+++ src/main/java/com/eppo/sdk/dto/Allocation.java
@@ -6,6 +6,6 @@ import lombok.Data;
 
 @Data
 public class Allocation {
-  private float percentExposure;
+  private double percentExposure;
   private List<Variation> variations;
 }
